i dont know what this is

+++ cooking/src/main/java/moe/quill/nadare/cooking/temperature/TempHandler.kt
@@ -3,18 +3,18 @@ package moe.quill.nadare.cooking.temperature
 import java.util.*
 
 class TempHandler {
-    private val freezeStates = mutableMapOf<UUID, Long>()
-    private val maxTicks = 140L
+    private val freezeStates = mutableMapOf<UUID, Int>()
+    private val maxTicks = 140
 
-    fun setTicks(uuid: UUID, amount: Long) {
+    fun setTicks(uuid: UUID, amount: Int) {
         freezeStates.getOrPut(uuid) { amount.coerceAtMost(maxTicks).coerceAtLeast(0) }
     }
 
-    fun modifyTicks(uuid: UUID, amount: Long) {
+    fun modifyTicks(uuid: UUID, amount: Int) {
         setTicks(uuid, getTicks(uuid) + amount)
     }
 
-    fun getTicks(uuid: UUID): Long {
+    fun getTicks(uuid: UUID): Int {
         return freezeStates.getOrDefault(uuid, 0)
     }
 }
